ResortApp/requirements_production.txt
# FastAPI Core and ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database ORM and Migrations
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and Security
bcrypt==3.2.2
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
PyJWT==2.8.0
cryptography==41.0.7

# Data Validation and Serialization
pydantic==2.5.0
pydantic-core==2.14.1
email-validator==2.1.0
python-dateutil==2.8.2

# Environment Configuration
python-dotenv==1.0.0

# HTTP Client and CORS
httpx==0.25.2
starlette==0.27.0

# File Processing and Upload
aiofiles==23.2.1
Pillow==10.1.0

# Template Engine
jinja2==3.1.2

# Utilities and Type Hints
typing-extensions==4.8.0
annotated-types==0.6.0

# Background Tasks and Caching
redis==5.0.1

# Data Processing and Export
pandas==2.1.4
openpyxl==3.1.2

# PDF Generation
reportlab==4.0.7

# Logging and Monitoring
structlog==23.2.0

# Core Dependencies (from working requirements)
anyio==4.11.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
h11==0.16.0
idna==3.11
pyasn1==0.6.1
pycparser==2.23
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2

# Production Optimizations
gevent==23.9.1

# Error Tracking
sentry-sdk[fastapi]==1.38.0

# Health Checks
requests==2.31.0

# Phone Number Validation
phonenumbers==8.13.25

# QR Code Generation
qrcode[pil]==7.4.2

# JSON Processing
orjson==3.9.10

# Payment Integration
stripe==7.7.0

# Email Services
fastapi-mail==1.4.1

# Rate Limiting
slowapi==0.1.9

# Timezone Support
pytz==2023.3

# System Monitoring
psutil==5.9.6

# Configuration Management
pydantic-settings==2.1.0

# WebSocket Support
websockets==12.0

# Backup and Cloud Storage
boto3==1.34.0

# Two-Factor Authentication
pyotp==2.9.0

# Scheduling
APScheduler==3.10.4

# Testing (Optional for production)
pytest==7.4.3
pytest-asyncio==0.21.1
